#myAccountTriggerLink --> no longer a <a>
The whole of #myAccountTriggerLink text is no longer an anchor <1>, but
the Sign In text still is.  This was done to allow some users who are
on a tablet but still qualify for the desktop version of the menu based
on the wide screen of their landscape tablet view, to be able to
trigger the menu dropdown by touching the #myAccountTriggerLink and not
triggering the Sign In function unless the actual text is touched.